refactor: split hero detection logic
