+++ packages/clip-processor-py/src/__init__.py
@@ -9,6 +9,7 @@ to extract Dota 2 hero information.
 try:
     from .clip_utils import get_clip_details, download_clip, download_single_frame, extract_frames
     from .dota_hero_detection import process_clip_url, process_stream_username, load_heroes_data
+    from .detection.image_processing import process_frame_for_heroes
     from .dota_heroes import get_hero_data
     from .api_server import app
 except ImportError:
